move model_register.py to omigami/spectra_matching/ and move its test file accordingly
--- don't change below this line ---
MLOPS-449 #time 59m
