+++ omigami/test/spectra_matching/test_mlflow_register_model.py
@@ -6,7 +6,7 @@ import pytest
 from mlflow.pyfunc import PythonModel
 
 from omigami.config import CODE_PATH
-from omigami.model_register import MLFlowModelRegister
+from omigami.spectra_matching.model_register import MLFlowModelRegister
 from omigami.spectra_matching.spec2vec.predictor import Spec2VecPredictor
 
 os.chdir(Path(__file__).parents[2])
